+++ src/main/java/com/devbliss/gpullr/controller/dto/UserConverter.java
@@ -1,6 +1,7 @@
 package com.devbliss.gpullr.controller.dto;
 
 import com.devbliss.gpullr.domain.User;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 /**
@@ -9,11 +10,18 @@ import org.springframework.stereotype.Component;
 @Component
 public class UserConverter {
 
+  @Autowired
+  private UserSettingsConverter userSettingsConverter;
+
   public UserDto toDto(User entity) {
     UserDto userDto = new UserDto();
     userDto.id = entity.id;
     userDto.username = entity.username;
     userDto.avatarUrl = entity.avatarUrl;
+
+    if (entity.userSettings != null) {
+      userDto.userSettingsDto = userSettingsConverter.toDto(entity.userSettings);
+    }
 
     return userDto;
   }