+++ src/main/java/com/devbliss/gpullr/controller/dto/UserSettingsDto.java
@@ -1,0 +1,17 @@
+package com.devbliss.gpullr.controller.dto;
+
+/**
+ * DTO for JSON de-/serialization of {@link com.devbliss.gpullr.domain.UserSettings} instances.
+ */
+public class UserSettingsDto {
+
+  public enum OrderOptionDto {
+    ASC,
+    DESC
+  }
+
+  public long id;
+
+  public OrderOptionDto orderOptionDto;
+
+}