+++ src/main/java/com/devbliss/gpullr/domain/User.java
@@ -1,8 +1,12 @@
 package com.devbliss.gpullr.domain;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
 import javax.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.NotBlank;
 
@@ -26,17 +30,30 @@ public class User {
 
   public Boolean canLogin = false;
 
+  @OneToOne(
+      fetch = FetchType.EAGER,
+      targetEntity = UserSettings.class,
+      cascade = CascadeType.PERSIST,
+      orphanRemoval = true)
+  @JoinColumn(name = "user_settings_id")
+  public UserSettings userSettings;
+
   public User() {}
 
   public User(Integer id, String username, String avatarUrl) {
-    this(id, username, avatarUrl, false);
+    this(id, username, avatarUrl, false, null);
   }
 
   public User(Integer id, String username, String avatarUrl, boolean canLogin) {
+    this(id, username, avatarUrl, canLogin, null);
+  }
+
+  public User(Integer id, String username, String avatarUrl, Boolean canLogin, UserSettings userSettings) {
     this.id = id;
     this.username = username;
     this.avatarUrl = avatarUrl;
     this.canLogin = canLogin;
+    this.userSettings = userSettings;
   }
 
   @Override