+++ src/test/java/com/devbliss/gpullr/service/PullRequestServiceUnitTest.java
@@ -2,6 +2,7 @@ package com.devbliss.gpullr.service;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -11,6 +12,9 @@ import com.devbliss.gpullr.repository.PullRequestRepository;
 import com.devbliss.gpullr.repository.UserRepository;
 import com.devbliss.gpullr.service.github.GithubApi;
 import java.time.ZonedDateTime;
+import java.time.temporal.ChronoUnit;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 import org.junit.Before;
 import org.junit.Test;
@@ -22,9 +26,8 @@ import org.mockito.runners.MockitoJUnitRunner;
 
 /**
  * Tests certain behavior of {@link PullRequestService} with mocked dependencies.
- * 
- * @author Henning Schütz <henning.schuetz@devbliss.com>
  *
+ * @author Henning Schütz <henning.schuetz@devbliss.com>
  */
 @RunWith(MockitoJUnitRunner.class)
 public class PullRequestServiceUnitTest {
@@ -37,6 +40,10 @@ public class PullRequestServiceUnitTest {
 
   private static final Integer ANOTHER_ASSIGNEE_ID = 18;
 
+  private static final String OLD_PR_TITLE = "old";
+
+  private static final String NEW_PR_TITLE = "new";
+
   @Mock
   private GithubApi githubApi;
 
@@ -45,6 +52,9 @@ public class PullRequestServiceUnitTest {
 
   @Mock
   private UserRepository userRepository;
+
+  @Mock
+  private UserService userService;
 
   private User assignee;
 
@@ -68,12 +78,13 @@ public class PullRequestServiceUnitTest {
     assignee.id = ASSIGNEE_ID;
     anotherAssignee = new User();
     anotherAssignee.id = ANOTHER_ASSIGNEE_ID;
-    pullRequestService = new PullRequestService(pullRequestRepository, userRepository, githubApi);
+    pullRequestService = new PullRequestService(pullRequestRepository, userRepository, githubApi, userService);
     pullRequestFromLocalStorage = new PullRequest();
     pullRequestFromLocalStorage.id = ID;
     pullRequestFromGitHub = new PullRequest();
     pullRequestFromGitHub.id = ID;
     pullRequestFromGitHub.author = author;
+    pullRequestFromGitHub.createdAt = ZonedDateTime.now();
   }
 
   @Test
@@ -98,5 +109,31 @@ public class PullRequestServiceUnitTest {
     verify(pullRequestRepository).save(pullRequestCaptor.capture());
     assertEquals(anotherAssignee, pullRequestCaptor.getValue().assignee);
     assertTrue(IN_THE_PAST.isBefore(pullRequestCaptor.getValue().assignedAt));
+  }
+
+  @Test
+  public void findAllOpenRegardsUserOrderOptions() {
+    mockFindOpenPullRequests();
+    when(userService.whoAmI()).thenReturn(author);
+
+    List<PullRequest> allOpen = pullRequestService.findAllOpen();
+
+    verify(userService).whoAmI();
+    assertEquals(NEW_PR_TITLE, allOpen.get(0).title);
+  }
+
+  private void mockFindOpenPullRequests() {
+    PullRequest oldPr = new PullRequest();
+    oldPr.id = 1;
+    oldPr.title = OLD_PR_TITLE;
+    oldPr.createdAt = ZonedDateTime.now().minus(5, ChronoUnit.HOURS);
+
+    PullRequest newPr = new PullRequest();
+    newPr.id = 2;
+    newPr.title = NEW_PR_TITLE;
+    newPr.createdAt = ZonedDateTime.now();
+
+    when(pullRequestRepository.findAllByState(eq(PullRequest.State.OPEN)))
+      .thenReturn(Arrays.asList(newPr, oldPr));
   }
 }