+++ changelog.md
@@ -1,7 +1,9 @@
 
 ### Version 0.1.0-SNAPSHOT
 
+
  - [bug] correct rate limit exceeded / reset handling
+ - [feature] user settings for preferred sort order of pull requests
  - [feature] ranking calculation
  - [refactor] rolling logfiles per day and after 50MB
  - [feature] added assignedAt property to PullRequest Entity and Dto
@@ -24,5 +26,5 @@
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: hschuetz <henning.schuetz@devbliss.com> Fri Mar 13 13:48:43 CET 2015
+-- Last change by: Daniel Walldorf <daniel.walldorf@devbliss.com> Fri Mar 13 15:52:57 CET 2015